requirements.txt
absl-py==0.15.0
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.4
certifi==2022.9.24
charset-normalizer==2.0.12
clang==5.0
click==8.0.4
cmake==3.24.1.1
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.2
h5py==3.1.0
idna==3.4
importlib-metadata==4.8.3
keras==2.6.0
Keras-Preprocessing==1.1.2
loadenv==0.1.1
Markdown==3.3.7
numpy==1.19.5
oauthlib==3.2.1
opencv-python==4.6.0.66
opt-einsum==3.3.0
Pillow==8.4.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.20.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.15.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.12
Werkzeug==2.0.3
wrapt==1.12.1
zipp==3.6.0